fix: Fix psalm issues

+++ lib/Tools/Traits/TStringTools.php
@@ -167,21 +167,15 @@ trait TStringTools {
 		return implode('', $word);
 	}
 
-
-	/**
-	 * @param int $bytes
-	 *
-	 * @return string
-	 */
 	public function humanReadable(int $bytes): string {
 		if ($bytes == 0) {
 			return '0.00 B';
 		}
 
 		$s = ['B', 'KB', 'MB', 'GB', 'TB', 'PB'];
-		$e = floor(log($bytes, 1024));
+		$e = (int)floor(log($bytes, 1024));
 
-		return (string)round((float)$bytes / 1024 ** $e, 2) . ' ' . $s[$e];
+		return (string)round($bytes / (1024 ** $e), 2) . ' ' . $s[$e];
 	}
